flair-generator/requirements.txt
praw<4
praw-script-oauth>=0.1.1
Pillow>=3.3.0
